backend/requirements.txt
fastapi
httpx
python-dotenv
uvicorn
peft